swag-project/requirements.txt
Flask==3.1.0
azure-data-tables==12.7.0  
python-dotenv==1.1.0